[SS] feat: separate passport configurations
